feat: vibe-coded speech to text, dedicated component
