+++ apps/web/client/chat-textarea/chat-textarea.component.ts
@@ -1,6 +1,7 @@
 import { CodayEvent, InviteEvent } from '@coday/coday-events'
 import { CodayEventHandler } from '../utils/coday-event-handler'
 import { getPreference } from '../utils/preferences'
+import { SpeechToTextareaComponent } from './speech-to-textarea.component'
 
 export class ChatTextareaComponent implements CodayEventHandler {
   private chatForm: HTMLFormElement
@@ -14,11 +15,6 @@ export class ChatTextareaComponent implements CodayEventHandler {
   private promptHistory: string[] = []
   private historyIndex: number = -1
   private tempInput: string = ''
-
-  // Voice input properties
-  private recognition: any | null = null
-  private voiceButton: HTMLButtonElement | null = null
-  private isRecording: boolean = false
 
   /**
    * Detect operating system for keyboard shortcuts
@@ -54,6 +50,8 @@ export class ChatTextareaComponent implements CodayEventHandler {
     this.chatTextarea = document.getElementById('chat-input') as HTMLTextAreaElement
     this.chatLabel = document.getElementById('chat-label') as HTMLLabelElement
     this.submitButton = document.getElementById('send-button') as HTMLButtonElement
+
+    new SpeechToTextareaComponent(this.chatTextarea, this.submitButton)
 
     this.chatForm.onsubmit = async (event) => {
       event.preventDefault()
@@ -100,19 +98,11 @@ export class ChatTextareaComponent implements CodayEventHandler {
     // Update the button label initially
     this.updateSendButtonLabel()
 
-    // Initialize voice input
-    this.initializeVoiceInput()
-
     // Listen for preference changes
     window.addEventListener('storage', (event) => {
       if (event.key === 'coday-preferences') {
         this.updateSendButtonLabel()
       }
-    })
-
-    // Listen for voice language changes from options panel
-    window.addEventListener('voiceLanguageChanged', (event: any) => {
-      this.updateRecognitionLanguage()
     })
   }
 
@@ -236,241 +226,5 @@ export class ChatTextareaComponent implements CodayEventHandler {
     } catch (error) {
       console.error('Error occurred while sending message:', error)
     }
-  }
-
-  private initializeVoiceInput(): void {
-    // Check if Speech Recognition is available
-    const SpeechRecognition = (window as any).SpeechRecognition || (window as any).webkitSpeechRecognition
-    console.log('speechRecognition', SpeechRecognition)
-    if (!SpeechRecognition) return
-
-    // Create recognition instance
-    this.recognition = new SpeechRecognition()
-    this.recognition.continuous = true // Permet des phrases longues sans coupure
-    this.recognition.interimResults = true
-    this.recognition.maxAlternatives = 1
-    this.recognition.lang = this.getSelectedLanguage() // Utilise la langue sélectionnée
-
-    // Paramètres avancés pour améliorer la capture de phrases longues
-    // Ces paramètres ne sont pas standards mais supportés par certains navigateurs
-    try {
-      // Augmente le délai avant que la reconnaissance s'arrête automatiquement
-      if ('grammars' in this.recognition) {
-        // Certains navigateurs supportent des paramètres étendus
-        this.recognition.serviceURI = undefined // Force l'utilisation du service par défaut
-      }
-    } catch (e) {
-      // Les paramètres avancés ne sont pas supportés, on continue avec la config de base
-      console.log('Paramètres avancés de reconnaissance vocale non supportés')
-    }
-
-    // Handle results
-    this.recognition.onresult = (event: any) => {
-      let finalTranscript = ''
-
-      for (let i = event.resultIndex; i < event.results.length; i++) {
-        const transcript = event.results[i][0].transcript
-        if (event.results[i].isFinal) {
-          finalTranscript += transcript
-        }
-      }
-
-      if (finalTranscript) {
-        // Post-traitement pour améliorer la ponctuation
-        const processedTranscript = this.improveTranscriptPunctuation(finalTranscript.trim())
-        this.appendToTextarea(processedTranscript)
-      }
-    }
-
-    this.recognition.onend = () => {
-      console.log('Recognition ended. Was recording:', this.isRecording)
-      this.isRecording = false
-      this.updateVoiceButtonState()
-    }
-
-    this.recognition.onerror = (event: any) => {
-      console.error('Speech recognition error:', event.error)
-      console.log('Error details:', {
-        error: event.error,
-        message: event.message,
-        timeStamp: event.timeStamp,
-      })
-      this.isRecording = false
-      this.updateVoiceButtonState()
-    }
-
-    // Événement pour déboguer les démarrages/arrêts
-    this.recognition.onstart = () => {
-      console.log('Recognition started successfully')
-    }
-
-    this.recognition.onspeechstart = () => {
-      console.log('Speech detected')
-    }
-
-    this.recognition.onspeechend = () => {
-      console.log('Speech ended - this might cause automatic stop')
-    }
-
-    this.createVoiceControls()
-  }
-
-  private createVoiceControls(): void {
-    // Find the send button to place voice controls next to it
-    const buttonContainer = this.submitButton.parentElement
-    if (!buttonContainer) return
-
-    // Create voice button (no language selector here anymore)
-    this.voiceButton = document.createElement('button')
-    this.voiceButton.type = 'button'
-    this.voiceButton.className = 'voice-button'
-    this.voiceButton.tabIndex = 0 // Permet le focus au clavier
-    this.voiceButton.innerHTML = '🎤'
-    this.voiceButton.title = 'Maintenir clic souris, touch, ou barre espace pour parler'
-    this.voiceButton.style.marginRight = '8px'
-
-    // Insert before send button
-    buttonContainer.insertBefore(this.voiceButton, this.submitButton)
-
-    // Mouse events for push-to-talk
-    this.voiceButton.addEventListener('mousedown', (e) => {
-      e.preventDefault()
-      this.startRecording()
-    })
-
-    this.voiceButton.addEventListener('mouseup', () => {
-      this.stopRecording()
-    })
-
-    this.voiceButton.addEventListener('mouseleave', () => {
-      this.stopRecording()
-    })
-
-    // Touch events for mobile
-    this.voiceButton.addEventListener('touchstart', (e) => {
-      e.preventDefault()
-      this.startRecording()
-    })
-
-    this.voiceButton.addEventListener('touchend', (e) => {
-      e.preventDefault()
-      this.stopRecording()
-    })
-
-    // Keyboard accessibility: Space bar = push-to-talk when button has focus
-    this.voiceButton.addEventListener('keydown', (e: KeyboardEvent) => {
-      if (e.code === 'Space' || e.key === ' ' || e.keyCode === 32) {
-        e.preventDefault()
-        // Démarre l'enregistrement seulement si ce n'est pas déjà en cours
-        if (!this.isRecording) {
-          this.startRecording()
-        }
-      }
-    })
-
-    this.voiceButton.addEventListener('keyup', (e: KeyboardEvent) => {
-      if (e.code === 'Space' || e.key === ' ' || e.keyCode === 32) {
-        e.preventDefault()
-        this.stopRecording()
-      }
-    })
-  }
-
-  private async startRecording(): Promise<void> {
-    if (!this.recognition || this.isRecording) return
-
-    try {
-      this.isRecording = true
-      this.updateVoiceButtonState()
-      this.recognition.start()
-    } catch (error) {
-      console.error('Failed to start recording:', error)
-      this.isRecording = false
-      this.updateVoiceButtonState()
-    }
-  }
-
-  private stopRecording(): void {
-    if (!this.recognition || !this.isRecording) {
-      this.appendToTextarea('\n\n') // add paragraph separation
-      return
-    }
-
-    try {
-      this.recognition.stop()
-    } catch (error) {
-      console.error('Failed to stop recording:', error)
-    }
-  }
-
-  private updateVoiceButtonState(): void {
-    if (!this.voiceButton) return
-
-    if (this.isRecording) {
-      this.voiceButton.classList.add('recording')
-      this.voiceButton.innerHTML = '🔴'
-      this.voiceButton.title = 'Recording... (release to stop)'
-    } else {
-      this.voiceButton.classList.remove('recording')
-      this.voiceButton.innerHTML = '🎤'
-      this.voiceButton.title = 'Maintenir clic souris, touch, ou barre espace pour parler'
-    }
-  }
-
-  /**
-   * Récupère la langue sélectionnée depuis les préférences globales
-   */
-  private getSelectedLanguage(): string {
-    return getPreference<string>('voiceLanguage', 'en-US') ?? 'en-US'
-  }
-
-  /**
-   * Met à jour la langue de reconnaissance vocale
-   */
-  private updateRecognitionLanguage(): void {
-    if (this.recognition) {
-      const selectedLang = this.getSelectedLanguage()
-      this.recognition.lang = selectedLang
-      console.log('Language changed to:', selectedLang)
-    }
-  }
-
-  /**
-   * Améliore la ponctuation du texte transcrit
-   */
-  private improveTranscriptPunctuation(text: string): string {
-    let improved = text
-
-    // Ajoute un point à la fin si pas de ponctuation
-    if (!/[.!?]$/.test(improved.trim())) {
-      improved = improved.trim() + '.'
-    }
-
-    // Capitalise la première lettre
-    improved = improved.charAt(0).toUpperCase() + improved.slice(1)
-
-    // Remplace certains mots-clés par de la ponctuation (optionnel)
-    improved = improved.replace(/\bpoint\b/gi, '.')
-    improved = improved.replace(/\bvirgule\b/gi, ',')
-    improved = improved.replace(/\bpoint d'interrogation\b/gi, '?')
-    improved = improved.replace(/\bpoint d'exclamation\b/gi, '!')
-    improved = improved.replace(/\bdeux points\b/gi, ':')
-    improved = improved.replace(/\bpoint virgule\b/gi, ';')
-
-    // Nettoie les espaces multiples
-    improved = improved.replace(/\s+/g, ' ')
-
-    return improved
-  }
-
-  private appendToTextarea(text: string): void {
-    const currentValue = this.chatTextarea.value
-    const newValue = currentValue ? `${currentValue} ${text}` : text
-    this.chatTextarea.value = newValue
-    this.chatTextarea.focus()
-
-    // Move cursor to end
-    const length = this.chatTextarea.value.length
-    this.chatTextarea.setSelectionRange(length, length)
   }
 }
